Factor out reference counter.
